requirements.txt
flask==2.3.2
pymongo==4.3.3
python-dotenv==1.0.0
tensorflow==2.10.0
gitpython==3.1.31
gdown>=4.6.0
requests
keras
numpy